wip: platform/win32
