+++ foo_input_ncm.vcxproj
@@ -90,14 +90,14 @@
       <ConformanceMode>true</ConformanceMode>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <LanguageStandard_C>stdc11</LanguageStandard_C>
-      <AdditionalIncludeDirectories>src; vendor; vendor/SDK</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>src; vendor; vendor/sdk</AdditionalIncludeDirectories>
       <AdditionalOptions>/Zc:__cplusplus %(AdditionalOptions)</AdditionalOptions>
       <PrecompiledHeader>Use</PrecompiledHeader>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/SDK/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
+      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/sdk/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
@@ -110,7 +110,7 @@
       <ConformanceMode>true</ConformanceMode>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <LanguageStandard_C>stdc11</LanguageStandard_C>
-      <AdditionalIncludeDirectories>src; vendor; vendor/SDK</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>src; vendor; vendor/sdk</AdditionalIncludeDirectories>
       <AdditionalOptions>/Zc:__cplusplus %(AdditionalOptions)</AdditionalOptions>
       <PrecompiledHeader>Use</PrecompiledHeader>
     </ClCompile>
@@ -119,7 +119,7 @@
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/SDK/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
+      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/sdk/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
@@ -130,14 +130,14 @@
       <ConformanceMode>true</ConformanceMode>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <LanguageStandard_C>stdc11</LanguageStandard_C>
-      <AdditionalIncludeDirectories>src; vendor; vendor/SDK</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>src; vendor; vendor/sdk</AdditionalIncludeDirectories>
       <AdditionalOptions>/Zc:__cplusplus %(AdditionalOptions)</AdditionalOptions>
       <PrecompiledHeader>Use</PrecompiledHeader>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/SDK/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
+      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/sdk/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
@@ -150,7 +150,7 @@
       <ConformanceMode>true</ConformanceMode>
       <LanguageStandard>stdcpp20</LanguageStandard>
       <LanguageStandard_C>stdc11</LanguageStandard_C>
-      <AdditionalIncludeDirectories>src; vendor; vendor/SDK</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>src; vendor; vendor/sdk</AdditionalIncludeDirectories>
       <AdditionalOptions>/Zc:__cplusplus %(AdditionalOptions)</AdditionalOptions>
       <PrecompiledHeader>Use</PrecompiledHeader>
     </ClCompile>
@@ -159,16 +159,17 @@
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/SDK/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
+      <AdditionalDependencies>$(CoreLibraryDependencies);%(AdditionalDependencies);vendor/sdk/foobar2000/shared/shared-$(Platform).lib</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="src\album_art_extractor.hpp" />
     <ClInclude Include="src\cipher\abnormal_RC4.hpp" />
-    <ClInclude Include="src\cipher\aes.h" />
+    <ClInclude Include="src\cipher\aes.hpp" />
     <ClInclude Include="src\cipher\aes_win32.hpp" />
     <ClInclude Include="src\cipher\cipher.h" />
     <ClInclude Include="src\cipher\exception.hpp" />
+    <ClInclude Include="src\cipher\utils.hpp" />
     <ClInclude Include="src\common\consts.hpp" />
     <ClInclude Include="src\common\helpers.hpp" />
     <ClInclude Include="src\common\platform.hpp" />
@@ -179,7 +180,7 @@
   <ItemGroup>
     <ClCompile Include="src\album_art_extractor.cpp" />
     <ClCompile Include="src\cipher\abnormal_RC4.cpp" />
-    <ClCompile Include="src\cipher\aes.cpp" />
+    <ClCompile Include="src\cipher\aes_win32.cpp" />
     <ClCompile Include="src\input_ncm.cpp" />
     <ClCompile Include="src\main.cpp" />
     <ClCompile Include="src\ncm_file.cpp" />
