+++ src/cipher/cipher.h
@@ -1,4 +1,5 @@
 #pragma once
 #include "exception.hpp"
 #include "abnormal_RC4.hpp"
-#include "aes.h"
+#include "aes.hpp"
+#include "utils.hpp"
